analytics/requirements.txt
fastapi
uvicorn
pymongo
pandas
scikit-learn
mlflow
python-dotenv
httpx
